requirements.txt
scikit-learn
pandas
kafka-python
catboost
peewee
pydantic
playhouse
pymysql
